requirements.txt
fastapi==0.115.5
langchain-core==0.3.31
langchain-google-genai==2.0.6
numpy==1.26.4
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.32.1